@@ -1,24 +1,21 @@
 @page "/myorders/{orderId:int}"
 @inject NavigationManager NavigationManager
 @inject HttpClient HttpClient
-
+@implements IDisposable
 
 <div class="main">
     @if (invalidOrder)
     {
         <h2>Order not found</h2>
-        <p>We;re sorry but this order no longer exists.</p>
+        <p>We're sorry but this order no longer exists.</p>
     }
-    else if (orderWithStatus == null)
+    else if (orderWithStatus is null)
     {
         <div class="track-order">
             <div class="track-order-title">
                 <h2>
                     <text>Loading...</text>
                 </h2>
-                <p class="ml-auto mb-0">
-                    ...
-                </p>
             </div>
         </div>
     }
@@ -35,12 +32,18 @@
             </div>
             <div class="track-order-body">
                 <div class="track-order-details">
+                    @if (IsOrderIncomplete)
+                    {
+                        <div class="spinner-grow text-danger float-right" role="status">
+                            <span class="sr-only">Checking your order status...</span>
+                        </div>
+                    }
                     @foreach (var pizza in orderWithStatus.Order.Pizzas)
                     {
                         <p>
                             <strong>
                                 @(pizza.Size)"
-                                @pizza.Special.Name
+                                @pizza.Special!.Name
                                 (£@pizza.GetFormattedTotalPrice())
                             </strong>
                         </p>
@@ -53,13 +56,24 @@
 
 @code {
     [Parameter] public int OrderId { get; set; }
-    OrderWithStatus orderWithStatus;
+
+    OrderWithStatus? orderWithStatus;
     bool invalidOrder = false;
-    protected override async Task OnParametersSetAsync()
+    bool IsOrderIncomplete =>
+    orderWithStatus is null || orderWithStatus.IsDelivered == false;
+
+    PeriodicTimer timer = new(TimeSpan.FromSeconds(3));
+    protected override async Task OnParametersSetAsync() =>
+    await GetLatestOrderStatusUpdatesAsync();
+    protected override Task OnAfterRenderAsync(bool firstRender) =>
+    firstRender ? StartPollingTimerAsync() : Task.CompletedTask;
+
+    async Task GetLatestOrderStatusUpdatesAsync()
     {
         try
         {
-            orderWithStatus = await HttpClient.GetFromJsonAsync<OrderWithStatus>($"{NavigationManager.BaseUri}orders/{OrderId}");
+            orderWithStatus = await HttpClient.GetFromJsonAsync<OrderWithStatus>(
+            $"{NavigationManager.BaseUri}orders/{OrderId}");
         }
         catch (Exception ex)
         {
@@ -67,4 +81,15 @@
             Console.Error.WriteLine(ex);
         }
     }
+
+    async Task StartPollingTimerAsync()
+    {
+        while (IsOrderIncomplete && await timer.WaitForNextTickAsync())
+        {
+            await GetLatestOrderStatusUpdatesAsync();
+            StateHasChanged();
+        }
+    }
+
+    public void Dispose() => timer.Dispose();
 }
