@@ -2,42 +2,34 @@ namespace BlazingPizza.Services;
 
 public class OrderState
 {
-    public bool ShowingConfigureDialog { get; private set; }
-    public Pizza ConfiguringPizza { get; private set; }
-    public Order Order { get; private set; } = new Order();
+    public bool ShowingConfigureDialog => ConfiguringPizza is not null;
+    public Pizza? ConfiguringPizza { get; private set; }
+    public Order Order { get; private set; } = new();
 
     public void ShowConfigurePizzaDialog(PizzaSpecial special)
     {
-        ConfiguringPizza = new Pizza()
+        ConfiguringPizza = new()
         {
             Special = special,
             SpecialId = special.Id,
             Size = Pizza.DefaultSize,
-            Toppings = new List<PizzaTopping>()
+            Toppings = [],
         };
-        ShowingConfigureDialog = true;
     }
 
     public void CancelConfigurePizzaDialog()
     {
         ConfiguringPizza = null;
-        ShowingConfigureDialog = false;
     }
 
     public void ConfirmConfigurePizzaDialog()
     {
-        Order.Pizzas.Add(ConfiguringPizza);
+        Order.Pizzas.Add(ConfiguringPizza!);
         ConfiguringPizza = null;
-        ShowingConfigureDialog = false;
     }
 
-    public void RemoveConfiguredPizza(Pizza pizza)
-    {
+    public void RemoveConfiguredPizza(Pizza pizza) =>
         Order.Pizzas.Remove(pizza);
-    }
 
-    public void ResetOrder()
-    {
-        Order = new Order();
-    }
+    public void ResetOrder() => Order = new();
 }
