@@ -10,22 +10,24 @@
     <EditForm EditContext=editContext OnValidSubmit=PlaceOrder>
         <div class="checkout-cols">
             <div class="checkout-order-details">
-                <h4>Review Order</h4>
+                <h4>Review order</h4>
                 <OrderReview Order="Order" />
             </div>
-            <div class="checkout-delivery=address">
+
+            <div class="checkout-delivery-address">
                 <h4>Deliver to...</h4>
                 @if (isError)
                 {
-                    <div class="alert alert-danger">Please enter a name and address.</div>
+                    <div class="alert alert-danger">Please enter a vaild name and address.</div>
                 }
                 <AddressEditor Address="Order.DeliveryAddress" />
             </div>
-
         </div>
+
         <button class="checkout-button btn btn-warning" type="Submit" disabled=@isError>
             Place order
         </button>
+
         <DataAnnotationsValidator />
     </EditForm>
 </div>
