@@ -1,24 +1,22 @@
 @page "/"
+@layout MainLayout
 @inject HttpClient HttpClient
 @inject NavigationManager NavigationManager
-@using BlazingPizza.Services
 @inject OrderState OrderState
+@inject IJSRuntime JavaScript
 
 <div class="main">
     <ul class="pizza-cards">
-        @if (specials != null)
+        @foreach (var special in specials ?? new())
         {
-            @foreach (var special in specials)
-            {
-                <li @onclick="@(()=> OrderState.ShowConfigurePizzaDialog(special))"
-                    style="background-image: url('@special.ImageUrl')">
-                    <div class="pizza-info">
-                        <span class="title">@special.Name</span>
-                        @special.Description
-                        <span class="price">@special.GetFormattedBasePrice()</span>
-                    </div>
-                </li>
-            }
+            <li @onclick="@(() => OrderState.ShowConfigurePizzaDialog(special))"
+                style="background-image: url('@special.ImageUrl')">
+                <div class="pizza-info">
+                    <span class="title">@special.Name</span>
+                    @special.Description
+                    <span class="price">@special.GetFormattedBasePrice()</span>
+                </div>
+            </li>
         }
     </ul>
 </div>
@@ -29,46 +27,68 @@
         OnConfirm="OrderState.ConfirmConfigurePizzaDialog" />
 }
 
-<div class="sidebar">
+<div class="sidebar @(order.Pizzas.Any() ? "with-total" : "")">
     @if (order.Pizzas.Any())
     {
         <div class="order-contents">
             <h2>Your order</h2>
+
             @foreach (var configuredPizza in order.Pizzas)
             {
                 <div class="cart-item">
-                    <div class="title">@(configuredPizza.Size)" @configuredPizza.Special.Name</div>
+                    <button type="button" class="close text-danger" aria-label="Close"
+                        @onclick="@(async () => await RemovePizzaConfirmation(configuredPizza))">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                    <div class="title">@(configuredPizza.Size)" @configuredPizza.Special!.Name</div>
                     <div class="item-price">
                         @configuredPizza.GetFormattedTotalPrice()
                     </div>
-                    <a @onclick="@(()=> OrderState.RemoveConfiguredPizza(configuredPizza))" class="delete-item">x</a>
                 </div>
             }
         </div>
     }
     else
     {
-        <div class="empty-cart">
-            Choose a pizza<br>to get started
-        </div>
+        <div class="empty-cart">Choose a pizza<br>to get started</div>
     }
 
-    <div class="order-total @(order.Pizzas.Any() ? "":"hidden")">
+    <div class="order-total @(order.Pizzas.Any() ? "" : "hidden")">
         Total:
         <span class="total-price">@order.GetFormattedTotalPrice()</span>
-        <a href="checkout" class="@(OrderState.Order.Pizzas.Count == 0 ? "btn btn-warning disabled":"btn btn-warning")">
+        <a href="checkout"
+            class="@(OrderState.Order.Pizzas.Count is 0 ? "btn btn-warning disabled" : "btn btn-warning")">
             Order >
         </a>
     </div>
 </div>
 
 @code {
-    List<PizzaSpecial> specials = new();
+    List<PizzaSpecial>? specials = new();
     Order order => OrderState.Order;
 
-    protected override async Task OnInitializedAsync()
-    {
-        specials = await HttpClient.GetFromJsonAsync<List<PizzaSpecial>>(NavigationManager.BaseUri + "specials");
-    }
+    protected override async Task OnInitializedAsync() =>
+    specials = await HttpClient.GetFromJsonAsync<List<PizzaSpecial>>(
+    $"{NavigationManager.BaseUri}specials");
 
+    async Task RemovePizzaConfirmation(Pizza removePizza)
+    {
+        var messageParams = new
+        {
+            title = "Remove Pizza?",
+            text = $"""Do you want to remove the "{removePizza.Special!.Name}" from your order?""",
+            icon = "warning",
+            buttons = new
+            {
+                abort = new { text = "No, leave it in my order", value = false },
+                confirm = new { text = "Yes, remove pizza", value = true }
+            },
+            dangerMode = true
+        };
+
+        if (await JavaScript.InvokeAsync<bool>("swal", messageParams))
+        {
+            OrderState.RemoveConfiguredPizza(removePizza);
+        }
+    }
 }
