@@ -1,4 +1,3 @@
-using BlazingPizza.Data;
 using BlazingPizza.Services;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -9,31 +8,31 @@ builder.Services.AddHttpClient();
 builder.Services.AddSqlite<PizzaStoreContext>("Data Source=pizza.db");
 builder.Services.AddScoped<OrderState>();
 
-
 var app = builder.Build();
-
-if (!app.Environment.IsDevelopment())
-{
-    app.UseExceptionHandler("/Error");
-}
-
-app.UseStaticFiles();
-app.UseRouting();
-
-app.MapRazorPages();
-app.MapBlazorHub();
-app.MapFallbackToPage("/_Host");
-app.MapControllerRoute("default", "{controller=Home}/{action-Index}/{id?}");
 
 // Initialize the database
 var scopeFactory = app.Services.GetRequiredService<IServiceScopeFactory>();
 using (var scope = scopeFactory.CreateScope())
 {
     var db = scope.ServiceProvider.GetRequiredService<PizzaStoreContext>();
-    // If there isn't a database already created, create one
-    if (db.Database.EnsureCreated())
+    if (await db.Database.EnsureCreatedAsync())
     {
-        SeedData.Initialize(db);
+        await SeedData.InitializeAsync(db);
     }
 }
+
+if (!app.Environment.IsDevelopment())
+{
+    app.UseExceptionHandler("/Error");
+    app.UseHsts();
+
+}
+app.UseHttpsRedirection();
+app.UseStaticFiles();
+app.UseRouting();
+app.MapRazorPages();
+app.MapBlazorHub();
+app.MapFallbackToPage("/_Host");
+app.MapControllerRoute("default", "{controller=Home}/{action-Index}/{id?}");
+
 app.Run();
