@@ -43,5 +43,6 @@
 </div>
 
 @code {
-    [Parameter] public Address Address { get; set; }
+    [Parameter, EditorRequired]
+    public required Address Address { get; set; }
 }
